Refactoration into file disposition
- extension variation is normalized
- best test assertion of image manipulation
