+++ imageManipulation_test.go
@@ -12,38 +12,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func Test_defineFormat(t *testing.T) {
-	t.Run("ImageWithCorrectFormatShouldReturnCorrectType", func(t *testing.T) {
-		formats := map[string][]string{
-			"jpeg": {"ulfric-stormcloak.jpg", "general tullius.jpeg", "esbern.jpeg", "arngeir.jpg"},
-			"png":  {"guts.png", "griffith.png", "casca.png", "judeau.png", "puck.png"},
-		}
-		for _, formatType := range formats {
-			for _, fileName := range formatType {
-				_, _, _, err := fixImgExtension(fileName)
-				assert.Nil(t, err)
-			}
-		}
-	})
-
-	t.Run("InvalidFormatShouldThrowException", func(t *testing.T) {
-		invalidSource := "i.wrong"
-		img, _, _, err := fixImgExtension(invalidSource)
-		assert.NotNil(t, err)
-		assert.EqualError(t, errors.New("wrong is not a valid format"), err.Error())
-		assert.Empty(t, img)
-	})
-
-	t.Run("ShouldNotFormatMalformedImageSource", func(t *testing.T) {
-		malformerSource := "sdfs"
-		format, _, _, err := fixImgExtension(malformerSource)
-		assert.NotNil(t, err)
-		assert.EqualError(t, fmt.Errorf(
-			"%s is malformed", malformerSource), err.Error())
-		assert.Empty(t, format)
-	})
-}
-
 func setupGenerateImgsTest(t *testing.T, fileName, ext, mimeTyp string) (map[string]*bufMedia, error) {
 	b, err := ioutil.ReadFile(fmt.Sprintf("sample_image/%s.%s", fileName, ext))
 	assert.Nil(t, err)
