+++ imageWriter_test.go
